Updating recipes is done

+++ src/components/Recipe.vue
@@ -1,12 +1,15 @@
 <template>
-  <div class="recipe four wide column">
+  <div class="recipe eight wide column">
     <div class="ui centered card">
+      <!--Content for displaying recipe-->
+      <div class="image">
+        <img :src=recipe.image />
+      </div>
       <div class='content' v-show="!isEditing">
         <div class="header">
           {{ recipe.title }}
         </div>
         <div class='meta'>
-          <img :src=recipe.image />
           <h4>Ingredients:</h4>
           <Ingredient
             v-for="ingredient in ingredients"
@@ -19,10 +22,21 @@
         </div>
         <div class='extra content'>
           <span class='right floated edit icon' v-on:click="showForm">
-          <i class='edit icon' />
+          <button class="ui primary basic button top-margin-button">
+             <i class='edit icon' /> Edit
+          </button>
           </span>
         </div>
       </div>
+      <div
+        class='ui bottom attached red basic button'
+        v-show="!isEditing"
+        v-on:click="deleteRecipe"
+      >
+      <i class='trash alternate outline icon' />
+        Delete
+      </div>
+      <!--Content for updating recipe-->
       <div class="content" v-show="isEditing">
         <div class='ui form'>
           <div class='field'>
@@ -40,25 +54,27 @@
               v-bind:key="ingredient.name"
               v-bind:ingredient="ingredient"
               v-bind:isEditing="isEditing"
+              v-model="recipe.ingredients"
             />
+            <div class="center">
+              <button
+                class="ui secondary basic button top-margin-button"
+                v-on:click="addIngredient">
+                <i class='plus icon' /> Add
+              </button>
+            </div>
           </div>
           <div class='field'>
             <label>Instructions</label>
-            <input type='text' v-model="recipe.instructions" >
-          </div>
-          <div class='ui bottom attached button'>
-            <button class='ui green basic button' v-on:click="hideForm">
-              Close and save
-            </button>
+            <textarea type='text' v-model="recipe.instructions" />
           </div>
         </div>
       </div>
       <div
-        class='ui bottom attached red basic button'
-        v-show="!isEditing"
-        v-on:click="deleteRecipe"
-      >
-        Delete
+        class='ui bottom attached green basic button'
+        v-show="isEditing"
+        v-on:click="hideFormAndSave">
+        <i class='check icon' /> Save
       </div>
     </div>
   </div>
@@ -85,8 +101,20 @@ export default {
     showForm() {
       this.isEditing = true;
     },
-    hideForm() {
+    hideFormAndSave() {
       this.isEditing = false;
+      console.log(this.recipe);
+      axios.put('https://cors-anywhere.herokuapp.com/https://evening-peak-29761.herokuapp.com/recipes', this.recipe)
+        .then((response) => {
+          console.log(response);
+          swal({
+            type: 'success',
+            title: 'Your recipe has been saved',
+            showConfirmButton: false,
+            timer: 1500,
+          });
+        })
+        .catch(error => console.log(error));
     },
     deleteRecipe() {
       swal({
@@ -97,10 +125,68 @@ export default {
         confirmButtonColor: '#DD6B55',
         confirmButtonText: 'Yes, delete it!',
       }).then((result) => {
-        console.log(result);
-        axios.delete('https://cors-anywhere.herokuapp.com/https://evening-peak-29761.herokuapp.com/recipes', { data: this.recipe })
-          .then(response => swal('Deleted!', response.data.msg, 'success'))
-          .catch(error => console.log(error.response.data));
+        if (result.value) {
+          axios.delete('https://cors-anywhere.herokuapp.com/https://evening-peak-29761.herokuapp.com/recipes', { data: this.recipe })
+            .then((response) => {
+              this.$parent.deleteRecipe(this.recipe);
+              swal('Deleted!', response.data.msg, 'success');
+            })
+            .catch(error => console.log(error));
+        }
+      });
+    },
+    deleteIngredient(pIngredient) {
+      const idx = this.ingredients.indexOf(pIngredient);
+      this.ingredients.splice(idx, 1);
+    },
+    addIngredient() {
+      swal.mixin({
+        input: 'text',
+        inputAttributes: {
+          autocapitalize: 'on',
+        },
+        confirmButtonText: 'Next &rarr;',
+        showCancelButton: true,
+        progressSteps: ['1', '2', '3'],
+      }).queue([
+        {
+          title: 'Name',
+          text: 'Write just the name of the ingredient',
+        },
+        {
+          title: 'Quantity',
+          text: 'Write just the amount as a number',
+        },
+        {
+          title: 'Unit',
+          text: 'Write the unit of measurement (e.i. teaspoon, oz, lb)',
+        },
+      ]).then((result) => {
+        const ing = result.value;
+        if (ing) {
+          ing[1] = Number(ing[1]) ? Number(ing[1]) : 0;
+          const newIng = {
+            name: ing[0],
+            quantity: ing[1],
+            unit: ing[2],
+          };
+          const idx = this.ingredients.indexOf(newIng);
+          console.log(idx);
+          if (idx !== -1 || ing[0] === '') {
+            swal({
+              title: 'Oh no!',
+              html: `Either your ingredient: ${newIng.name} (${newIng.quantity} ${newIng.unit}) already exists or you forgot to write its name`,
+              confirmButtonText: 'Silly me!',
+            });
+          } else {
+            this.recipe.ingredients.push(newIng);
+            swal({
+              title: 'All done!',
+              html: `Your ingredient: ${newIng.name} (${newIng.quantity} ${newIng.unit}) was added`,
+              confirmButtonText: 'Yay!',
+            });
+          }
+        }
       });
     },
   },
@@ -110,5 +196,18 @@ export default {
 <style>
 .recipe {
   margin: 0;
+}
+.center{
+  text-align: center;
+}
+.top-margin-button{
+  margin-top: 15px !important;
+}
+.ui.centered.card{
+  width: 100%;
+}
+.image{
+  max-height: 500px !important;
+  overflow: hidden !important;
 }
 </style>
